+++ HISTORY.md
@@ -1,5 +1,5 @@
-unreleased
-==========
+1.4.0 / 2015-09-18
+==================
 
   * Accept array of secrets in addition to a single secret
   * Fix `JSONCookie` to return `undefined` for non-string arguments